+++ CHANGELOG.md
@@ -8,3 +8,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - new Benchmark test for mixed playback of 4k and FHD videos
 - new Video and Image and HTML example
+- added Capabilites examples